requirements.txt
accelerate==0.29.3
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.4
arff==0.9
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.1.0
asttokens==2.4.1
async-generator==1.10
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
azure-core==1.30.1
azure-storage-blob==12.19.1
Babel==2.14.0
backcall==0.2.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
beir==2.0.0
bitarray==2.8.1
bitstring==4.1.1
bleach==6.0.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.1.0
cryptography==39.0.1
cycler==0.11.0
Cython==0.29.36
dataclasses-json==0.6.5
datasets==2.19.0
deap==1.4.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
dnspython==2.4.2
ecdsa==0.18.0
elasticsearch==7.9.1
entrypoints==0.4
environs==9.5.0
et-xmlfile==1.1.0
evaluate==0.4.2
exceptiongroup==1.2.1
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.110.2
fastjsonschema==2.18.0
feedparser==6.0.10
ffmpy==0.3.2
filelock==3.14.0
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
funcsigs==1.0.2
future==1.0.0
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google_search_results==2.4.2
googleapis-common-protos==1.63.1
gradio==4.20.0
gradio_client==0.11.0
greenlet==2.0.2
grpcio==1.57.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.22.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.29.4
ipython==8.24.0
ipython-genutils==0.2.0
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.0
json5==0.9.20
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.2
jupyterlab==4.1.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.3
kiwisolver==1.4.5
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain-experimental==0.0.59
langchain-openai==0.1.8
langchain-text-splitters==0.2.0
langsmith==0.1.67
latexcodec==3.0.0
lfs==0.2
lockfile==0.12.2
loralib==0.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.2
matplotlib==3.7.2
matplotlib-inline==0.1.7
mdurl==0.1.2
minio==7.2.7
mistune==3.0.1
mock==5.1.0
monty==2024.5.24
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.3
nltk==3.8.1
nose==1.3.7
notebook_shim==0.2.4
numpy==1.25.2
oauthlib==3.2.2
openai==1.28.1
openpyxl==3.1.2
orjson==3.9.15
overrides==7.7.0
packaging==23.2
palettable==3.3.3
pandas==2.1.0
pandocfilters==1.5.0
paramiko==3.3.1
parso==0.8.4
path==16.7.1
path.py==12.5.0
pathlib2==2.3.7.post1
pbr==5.11.1
peft==0.10.0
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.9.1
plotly==5.22.0
prometheus-client==0.17.1
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybtex==0.24.0
pycparser==2.21
pycryptodome==3.10.1
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pymatgen==2024.5.31
pymilvus==2.4.1
PyNaCl==1.5.0
pyparsing==3.0.9
pypdf==4.1.0
pyperclip==1.8.2
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-youtube==0.9.4
pytrec_eval==0.5
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
rank-bm25==0.2.2
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.10.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.3.0
safetensors==0.4.1
scikit-learn==1.3.1
scipy==1.11.2
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers==2.5.0
sentencepiece==0.1.99
sgmllib3k==1.0.0
shellingham==1.5.4
simplegeneric==0.8.1
simplejson==3.19.2
singledispatch==4.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
spglib==2.4.0
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.37.2
streamlit==1.30.0
streamlit-aggrid==0.3.4.post3
streamlit-antd-components==0.3.2
streamlit-chatbox==1.1.11
streamlit-feedback==0.1.3
streamlit-modal==0.1.1
streamlit-option-menu==0.3.12
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
testpath==0.6.0
threadpoolctl==3.4.0
tiktoken==0.7.0
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.0
tornado==6.3.3
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.1
typer==0.9.4
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2023.3
tzlocal==5.2
ujson==5.7.0
uncertainties==3.1.7
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.29.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
websockets==11.0.3
wikipedia==1.4.0
xxhash==3.2.0
yarl==1.9.4
youtube-search==2.1.2
zipp==3.16.2
